requirements.txt
babel
deep-translator
django
forex-python
gunicorn
mapcode
money
mysqlclient
Pillow
python-dateutil
python-dotenv
python-http-client
pytz
requests
sqlparse
six
urllib3
virtualenv
zipp